message-board/requirements.txt
flask
psycopg2-binary
werkzeug
authlib
requests
python-dotenv